♻ refactor: 쿠폰 관리 컴포넌트 분리
